Improvements related to the ASOC_driver routines and the library method (A2E_LIB.py).

+++ kernel_A2E_MABU_aux.c
@@ -2,7 +2,7 @@
 
 void __kernel  split_absorbed(const    int    IDUST,
                               const    int    N,
-                              __global float *RABS,   // [ifreq, idust]
+                              __global double *RABS,   // [ifreq, idust]
                               __global float *ABU,    // [icell, idust]
                               __global float *IN,     // [N, NFREQ]
                               __global float *OUT     // [N, NFREQ]
@@ -14,5 +14,6 @@ void __kernel  split_absorbed(const    int    IDUST,
       den = 0.0f ;
       for(int idust=0; idust<NDUST; idust++)   den += ABU[icell*NDUST+idust]*RABS[ifreq*NDUST+idust] ;
       OUT[icell*NFREQ+ifreq] = IN[icell*NFREQ+ifreq] * RABS[ifreq*NDUST+IDUST] / den ;
+      // if (icell==10) printf(" %12.4e = %12.4e * %12.4e / % 12.4e\n",OUT[icell*NFREQ+ifreq], IN[icell*NFREQ+ifreq], RABS[ifreq*NDUST+IDUST], den) ;
    }
 }
